% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-census-tract-profiles.R
\name{get_census_tract_profiles}
\alias{get_census_tract_profiles}
\title{Get census profiles for all census tracts in a cma}
\usage{
get_census_tract_profiles(cma = NULL, topic = 9)
}
\arguments{
\item{topic}{}
}
\description{
Get census profiles for all census tracts in a cma
}
